Update voice and music plugins to use plugin base
