+++ TobysBot.Music/Commands/MusicPlugin.cs
@@ -1,0 +1,11 @@
+﻿using Discord.Commands;
+using TobysBot.Commands;
+
+namespace TobysBot.Music.Commands;
+
+[Name("Music")]
+[Summary("Listen to music on Discord with Toby's Bot")]
+public partial class MusicPlugin : PluginBase
+{
+    
+}
