+++ TobysBot.Voice/Commands/VoicePlugin.cs
@@ -1,0 +1,11 @@
+﻿using Discord.Commands;
+using TobysBot.Commands;
+
+namespace TobysBot.Voice.Commands;
+
+[Name("Voice")]
+[Summary("Manage Toby's Bot in the voice channel")]
+public partial class VoicePlugin : PluginBase
+{
+    
+}
